+++ bin/commands/BlastCommand.js
@@ -13,7 +13,7 @@ const BlastCommand = {
     },
 
     blast: async function(interaction, channel) {
-        this.targets = await this.findTargets(channel);
+        this.targets = await this.findTargets(channel, interaction);
         await interaction.reply({
             content: `${interaction.user} has blasted the entrance of the room.`,
             files: [{
@@ -26,19 +26,23 @@ const BlastCommand = {
         });
     },
     // TODO: You should not be able to blast yourself. Check that the user does not target themselves.
-    findTargets: async function(channel) {
+    findTargets: async function(channel, interaction) {
         const messages = await channel.messages.fetch({limit:50});
         // remove the first message because it will be the interaction
         let targets = [];
         // Iterate through the messages here with the variable "messages". Theoretically, the message should be a user or a bot interaction initiated by a user, a message by a user, or a message by a bot.
         messages.every(message => {
-            console.log(message.content)
+            // console.log(message.content)
             // Check if message author is a bot.
             if(message.author.bot) {
                 // If it is a bot, get the interaction user as the stand in.
-                if(message.interaction) targets.push(message.interaction.user);
+                // Check that the interaction.user.id !== message.interaction.user.id
+                // (interaction.user.id === message.interaction.user.id) ? console.log( 'Do nothing') : 
+                if(message.interaction) {
+                    if(interaction.user.id !== message.interaction.user.id) targets.push(message.interaction.user);
+                }
                
-            } else targets.push(message.member); // else the target is not a bot.
+            } else (interaction.user.id !== message.author.id) ? targets.push(message.member): console.log('Do nothing.'); // else the target is not a bot.
             
             // Check the size of the targets variable. If it is 3 in size, return false.
             if(targets.length >= 3) return false;